Relocated DirectChat functions to separate file
